Osetreni tajnych hlaseni a pripadu u prirazovani + oprava zobrazovani tajnych u prohlizeni a editace pripadu.

+++ trunk/BIStro/addreports.php
@@ -2,13 +2,17 @@
 	require_once ('./inc/func_main.php');
 	
 	if (isset($_POST['addtoareport'])) {
-		MySQL_Query ("DELETE FROM ".DB_PREFIX."ar2c WHERE ".DB_PREFIX."ar2c.idreport=".$_POST['reportid']);
+		if ($usrinfo['right_power']==1) {
+			MySQL_Query ("DELETE FROM ".DB_PREFIX."ar2c WHERE ".DB_PREFIX."ar2c.idreport=".$_POST['reportid']);
+		} else {
+			MySQL_Query ("DELETE c FROM ".DB_PREFIX."ar2c as c, ".DB_PREFIX."cases as p WHERE c.idcase=p.id AND p.secret=0 AND c.idreport=".$_POST['reportid']);
+		}
 		if (isset($_POST['case'])) {
 			$case=$_POST['case'];
 		}
 		pageStart ('Uložení změn');
 		mainMenu (5);
-		sparklets ('<a href="./reports.php">hlášení</a> &raquo; <a href="./editactrep.php?rid='.$_POST['reportid'].'">úprava hlášení</a> &raquo; <strong>uložení změn</strong>','<a href="readreport.php?rid='.$_POST['reportid'].'&hidenotes=0&truenames=0">zobrazit upravené</a>');
+		sparklets ('<a href="./reports.php">hlášení</a> &raquo; <a href="./editactrep.php?rid='.$_POST['reportid'].'">úprava hlášení</a> &raquo; <strong>uložení změn</strong>','<a href="readactrep.php?rid='.$_POST['reportid'].'&hidenotes=0&truenames=0">zobrazit upravené</a>');
 		echo '<div id="obsah"><p>Hlášení přiřazeno k příslušným případům.</p></div>';
 		if (isset($_POST['case'])) {
 			for ($i=0;$i<Count($case);$i++) {
@@ -19,7 +23,11 @@
 	}
 	
 	if (isset($_POST['addcasetoareport'])) {
-		MySQL_Query ("DELETE FROM ".DB_PREFIX."ar2c WHERE ".DB_PREFIX."ar2c.idcase=".$_POST['caseid']);
+		if ($usrinfo['right_power']==1) {
+			MySQL_Query ("DELETE FROM ".DB_PREFIX."ar2c WHERE ".DB_PREFIX."ar2c.idcase=".$_POST['caseid']);
+		} else {
+			MySQL_Query ("DELETE c FROM ".DB_PREFIX."ar2c as c, ".DB_PREFIX."reports as p WHERE c.idreport=p.id AND p.secret=0 AND c.idcase=".$_POST['caseid']);
+		}
 		if (isset($_POST['report'])) {
 			$report=$_POST['report'];
 		}
